JC: Split up steps
